Adding templates for decluttering
